third_party/WebKit/LayoutTests/fast/css/border-start-end-expected.txt

Tests border-start and border-end

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testWidth("ltr", "border-inline-start-width: 10px; border-inline-start-style: solid") is 110
PASS testWidth("ltr", "border-inline-end-width: 20px; border-inline-end-style:  solid") is 120
PASS testWidth("rtl", "border-inline-start-width: 10px; border-inline-start-style:  solid") is 110
PASS testWidth("rtl", "border-inline-end-width: 20px; border-inline-end-style:  solid") is 120
PASS test("ltr", "border-inline-start-color: rgb(255, 0, 0)", "border-left-color") is "rgb(255, 0, 0)"
PASS test("ltr", "border-left-color: rgb(255, 0, 0)", "border-inline-start-color") is "rgb(255, 0, 0)"
PASS test("ltr", "border-inline-end-color: rgb(255, 0, 0)", "border-right-color") is "rgb(255, 0, 0)"
PASS test("ltr", "border-right-color: rgb(255, 0, 0)", "border-inline-end-color") is "rgb(255, 0, 0)"
PASS test("rtl", "border-inline-start-color: rgb(255, 0, 0)", "border-right-color") is "rgb(255, 0, 0)"
PASS test("rtl", "border-right-color: rgb(255, 0, 0)", "border-inline-start-color") is "rgb(255, 0, 0)"
PASS test("rtl", "border-inline-end-color: rgb(255, 0, 0)", "border-left-color") is "rgb(255, 0, 0)"
PASS test("rtl", "border-left-color: rgb(255, 0, 0)", "border-inline-end-color") is "rgb(255, 0, 0)"
PASS test("ltr", "border-inline-start-style: dotted", "border-left-style") is "dotted"
PASS test("ltr", "border-left-style: dotted", "border-inline-start-style") is "dotted"
PASS test("ltr", "border-inline-end-style: dotted", "border-right-style") is "dotted"
PASS test("ltr", "border-right-style: dotted", "border-inline-end-style") is "dotted"
PASS test("rtl", "border-inline-start-style: dotted", "border-right-style") is "dotted"
PASS test("rtl", "border-right-style: dotted", "border-inline-start-style") is "dotted"
PASS test("rtl", "border-inline-end-style: dotted", "border-left-style") is "dotted"
PASS test("rtl", "border-left-style: dotted", "border-inline-end-style") is "dotted"
PASS test("ltr", "border-inline-start-width: 10px; border-inline-start-style: solid", "border-left-width") is "10px"
PASS test("ltr", "border-inline-end-width: 10px; border-inline-end-style: solid", "border-right-width") is "10px"
PASS test("rtl", "border-inline-start-width: 10px; border-inline-start-style: solid", "border-right-width") is "10px"
PASS test("rtl", "border-inline-end-width: 10px; border-inline-end-style: solid", "border-left-width") is "10px"
PASS test("ltr", "border-left: 10px solid", "border-inline-start-width") is "10px"
PASS test("ltr", "border-right: 10px solid", "border-inline-end-width") is "10px"
PASS test("rtl", "border-left: 10px solid", "border-inline-end-width") is "10px"
PASS test("rtl", "border-right: 10px solid", "border-inline-start-width") is "10px"
PASS test("ltr", "border-inline-start: 10px solid red", "border-left-color") is "rgb(255, 0, 0)"
PASS test("ltr", "border-inline-start: 10px solid red", "border-left-style") is "solid"
PASS test("ltr", "border-inline-start: 10px solid red", "border-left-width") is "10px"
PASS test("rtl", "border-inline-start: 10px solid red", "border-right-color") is "rgb(255, 0, 0)"
PASS test("rtl", "border-inline-start: 10px solid red", "border-right-style") is "solid"
PASS test("rtl", "border-inline-start: 10px solid red", "border-right-width") is "10px"
PASS test("ltr", "border-inline-end: 10px solid red", "border-right-color") is "rgb(255, 0, 0)"
PASS test("ltr", "border-inline-end: 10px solid red", "border-right-style") is "solid"
PASS test("ltr", "border-inline-end: 10px solid red", "border-right-width") is "10px"
PASS test("rtl", "border-inline-end: 10px solid red", "border-left-color") is "rgb(255, 0, 0)"
PASS test("rtl", "border-inline-end: 10px solid red", "border-left-style") is "solid"
PASS test("rtl", "border-inline-end: 10px solid red", "border-left-width") is "10px"
PASS successfullyParsed is true

TEST COMPLETE
